added tests fixed some compiler warnings

+++ MecabCocoa/darts.h
@@ -178,7 +178,7 @@ class DoubleArrayImpl {
                  static_cast<size_t>(siblings[siblings.size() - 1].code + 1));
 
     for (size_t i = 0; i < siblings.size(); ++i)
-      array_[begin + siblings[i].code].check = begin;
+      array_[begin + siblings[i].code].check = (int) begin;
 
     for (size_t i = 0; i < siblings.size(); ++i) {
       std::vector <node_t> new_siblings;
@@ -199,7 +199,7 @@ class DoubleArrayImpl {
 
       } else {
         size_t h = insert(new_siblings);
-        array_[begin + siblings[i].code].base = h;
+        array_[begin + siblings[i].code].base = (int) h;
       }
     }
